+++ src/L3-solution/BoSSS.Solution.LevelSetTools/FastMarching/NarrowMarchingBand.cs
@@ -998,7 +998,7 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
 
             MsrMatrix ExtVelMatrix;
             double[][] ExtVelRHS = new double[ComponentMode ? ExtVel.Length : 1][];
-            List<int> UsedRows = new List<int>();
+            List<long> UsedRows = new List<long>();
             {
                 var map = ExtVel[0].Mapping;
 
@@ -1127,17 +1127,17 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
                     int N = map.BasisS[0].GetLength(j);
                     for(int n = 0; n < N; n++) {
                         int iL = map.LocalUniqueCoordinateIndex(0, j, n);
-                        int iG = map.GlobalUniqueCoordinateIndex(0, j, n);
+                        long iG = map.GlobalUniqueCoordinateIndex(0, j, n);
 
                         if(CCbitmask[j]) {
                             //var Row = ExtVelMatrix.GetRow(iG);
                             //foreach (var entry in Row) {
                             int Lr;
-                            int[] row_cols = null;
+                            long[] row_cols = null;
                             double[] row_vals = null;
                             Lr = ExtVelMatrix.GetRow(iG, ref row_cols, ref row_vals);
                             for (int lr = 0; lr < Lr; lr++) {
-                                int ColIndex = row_cols[lr];
+                                long ColIndex = row_cols[lr];
                                 double Value = row_vals[lr];
                                 if (Value != 0.0) {
                                     int jL = map.Global2LocalIndex(ColIndex);
@@ -1158,9 +1158,9 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
                 }
 #endif
 
-                int i0 = map.i0;
-                int iE = map.iE;
-                for(int i = i0; i < iE; i++) {
+                long i0 = map.i0;
+                long iE = map.iE;
+                for(long i = i0; i < iE; i++) {
                     if(ExtVelMatrix.GetNoOfNonZerosPerRow(i) == 0) {
                         ExtVelMatrix[i, i] = 1.0;
                     } else {
@@ -1179,7 +1179,7 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
             MsrMatrix essExtVelMatrix;
             {
                 essExtVelMatrix = new MsrMatrix(new Partitioning(UsedRows.Count));
-                ExtVelMatrix.WriteSubMatrixTo(essExtVelMatrix, UsedRows, default(int[]), UsedRows, default(int[]));
+                ExtVelMatrix.WriteSubMatrixTo(essExtVelMatrix, UsedRows, default(long[]), UsedRows, default(long[]));
             }
 
 
@@ -1197,7 +1197,7 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
                     // extract rhs
                     double[] rhs = new double[UsedRows.Count];
                     double[] x = new double[UsedRows.Count];
-                    rhs.AccV(1.0, ExtVelRHS[d], default(int[]), UsedRows);
+                    rhs.AccV(1.0, ExtVelRHS[d], default(long[]), UsedRows);
 
                     double[] Resi_d = rhs.CloneAs();
                     double NORM_rhs = rhs.L2NormPow2().MPISum().Sqrt();
@@ -1216,7 +1216,7 @@ namespace BoSSS.Solution.LevelSetTools.Advection {
 
                     // write back solution
                     ExtVel[d].Clear();
-                    ExtVel[d].CoordinateVector.AccV(1.0, x, UsedRows, default(int[]));
+                    ExtVel[d].CoordinateVector.AccV(1.0, x, UsedRows, default(long[]));
                 }
             }
